grid2op.MakeEnv expose new Make implementation as default
