+++ grid2op/MakeEnv/MakeFromPath.py
@@ -1,0 +1,353 @@
+# Copyright (c) 2019-2020, RTE (https://www.rte-france.com)
+# See AUTHORS.txt
+# This Source Code Form is subject to the terms of the Mozilla Public License, version 2.0.
+# If a copy of the Mozilla Public License, version 2.0 was not distributed with this file,
+# you can obtain one at http://mozilla.org/MPL/2.0/.
+# SPDX-License-Identifier: MPL-2.0
+# This file is part of Grid2Op, Grid2Op a testbed platform to model sequential decision making in power systems.
+
+import os
+import importlib.util
+import pkg_resources
+import warnings
+import json
+
+from grid2op.Environment import Environment
+from grid2op.Backend import Backend, PandaPowerBackend
+from grid2op.Parameters import Parameters
+from grid2op.Chronics import ChronicsHandler, Multifolder, ChangeNothing
+from grid2op.Chronics import GridStateFromFile, GridStateFromFileWithForecasts, GridValue
+from grid2op.Action import BaseAction, TopologyAction, TopologyAndDispatchAction, DontAct
+from grid2op.Exceptions import *
+from grid2op.Observation import CompleteObservation, BaseObservation
+from grid2op.Reward import BaseReward, L2RPNReward, RedispReward
+from grid2op.Rules import BaseRules, AlwaysLegal, DefaultRules
+from grid2op.VoltageControler import ControlVoltageFromFile
+from grid2op.Opponent import BaseOpponent
+
+from grid2op.MakeEnv.get_default_aux import _get_default_aux
+
+ERR_MSG_KWARGS = {
+    "backend": "The backend of the environment (keyword \"backend\") must be an instance of grid2op.Backend",
+    "observation_class": "The type of observation of the environment (keyword \"observation_class\")" \
+    " must be a subclass of grid2op.BaseObservation",
+    "param": "The parameters of the environment (keyword \"param\") must be an instance of grid2op.Parameters",
+    "gamerules_class": "The type of rules of the environment (keyword \"gamerules_class\")" \
+    " must be a subclass of grid2op.BaseRules",
+    "reward_class": "The type of reward in the environment (keyword \"reward_class\") must be a subclass of grid2op.BaseReward",
+    "action_class": "The type of action of the environment (keyword \"action_class\") must be a subclass of grid2op.BaseAction",
+    "data_feeding_kwargs": "The argument to build the data generation process [chronics]" \
+    "  (keyword \"data_feeding_kwargs\") should be a dictionnary.",
+    "chronics_class": "The argument to build the data generation process [chronics] (keyword \"chronics_class\")" \
+    " should be a class that inherit grid2op.Chronics.GridValue.",
+    "chronics_handler": "The argument to build the data generation process [chronics] (keyword \"data_feeding\")" \
+    " should be a class that inherit grid2op.ChronicsHandler.ChronicsHandler.",
+    "voltagecontroler_class": "The argument to build the online controler for chronics (keyword \"volagecontroler_class\")" \
+    " should be a class that inherit grid2op.VoltageControler.ControlVoltageFromFile.",
+    "names_chronics_to_grid": "The converter between names (keyword \"names_chronics_to_backend\") should be a dictionnary.",
+    "other_rewards": "The argument to build the online controler for chronics (keyword \"other_rewards\") "
+                     "should be dictionnary.",
+    "opponent_action_class": "The argument used to build the \"opponent_action_class\" should be a class that "
+                             "inherit from \"BaseAction\"",
+    "opponent_class": "The argument used to build the \"opponent_class\" should be a class that "
+                             "inherit from \"BaseOpponent\"",
+    "opponent_init_budget": "The initial budget of the opponent \"opponent_init_budget\" should be a float",
+    "chronics_path": "The path where the data is located (keyword \"chronics_path\") should be a string.",
+    "grid_path": "The path where the grid is located (keyword \"grid_path\") should be a string."
+}
+
+NAME_CHRONICS_FOLDER = "chronics"
+NAME_GRID_FILE = "grid.json"
+NAME_GRID_LAYOUT_FILE = "grid_layout.json"
+NAME_CONFIG_FILE = "config.py"
+
+def _check_kwargs(kwargs):
+    for el in kwargs:
+        if not el in ERR_MSG_KWARGS.keys():
+            raise EnvError("Unknown keyword argument \"{}\" used to create an Environement. "
+                           "No Environment will be created. "
+                           "Accepted keyword arguments are {}".format(el, ERR_MSG_KWARGS.keys()))
+
+
+def _check_path(path, info):
+    if path is None or os.path.exists(path) is False:
+        raise EnvError("Cannot find {}. {}".format(path, info))
+
+
+def make_from_dataset_path(dataset_path="/", **kwargs):
+    """
+    This function is a shortcut to rapidly create environments within the grid2op Framework.
+
+    It mimic the ``gym.make`` function.
+
+    Parameters
+    ----------
+    dataset_path: ``str``
+        Path to the dataset folder
+
+    param: ``grid2op.Parameters.Parameters``, optional
+        Type of parameters used for the Environment. Parameters defines how the powergrid problem is cast into an
+        markov decision process, and some internal
+
+    backend: ``grid2op.Backend.Backend``, optional
+        The backend to use for the computation. If provided, it must be an instance of :class:`grid2op.Backend.Backend`.
+
+    action_class: ``type``, optional
+        Type of BaseAction the BaseAgent will be able to perform.
+        If provided, it must be a subclass of :class:`grid2op.BaseAction.BaseAction`
+
+    observation_class: ``type``, optional
+        Type of BaseObservation the BaseAgent will receive.
+        If provided, It must be a subclass of :class:`grid2op.BaseAction.BaseObservation`
+
+    reward_class: ``type``, optional
+        Type of reward signal the BaseAgent will receive.
+        If provided, It must be a subclass of :class:`grid2op.BaseReward.BaseReward`
+
+    gamerules_class: ``type``, optional
+        Type of "Rules" the BaseAgent need to comply with. Rules are here to model some operational constraints.
+        If provided, It must be a subclass of :class:`grid2op.RulesChecker.BaseRules`
+
+    data_feeding_kwargs: ``dict``, optional
+        Dictionnary that is used to build the `data_feeding` (chronics) objects.
+
+    chronics_class: ``type``, optional
+        The type of chronics that represents the dynamics of the Environment created. Usually they come from different
+        folders.
+
+    data_feeding: ``type``, optional
+        The type of chronics handler you want to use.
+
+    volagecontroler_class: ``type``, optional
+        The type of :class:`grid2op.VoltageControler.VoltageControler` to use, it defaults to
+
+    chronics_path: ``str``
+        Path where to look for the chronics dataset (optional)
+
+    grid_path: ``str``, optional
+        The path where the powergrid is located.
+        If provided it must be a string, and point to a valid file present on the hard drive.
+
+    Returns
+    -------
+    env: :class:`grid2op.Environment.Environment`
+        The created environment.
+    """
+    # Compute and find root folder
+    _check_path(dataset_path, "Dataset root directory")
+    dataset_path_abs = os.path.abspath(dataset_path)
+    # Compute env name from directory name
+    name_env = os.path.split(dataset_path_abs)[1]
+
+    # Compute and find chronics folder
+    chronics_path = _get_default_aux("chronics_path", kwargs,
+                                     defaultClassApp=str, defaultinstance='',
+                                     msg_error=ERR_MSG_KWARGS["chronics_path"])
+    if chronics_path == "":
+        # if no "chronics_path" argument is provided, look into the "chronics" folder
+        chronics_path_abs = os.path.abspath(os.path.join(dataset_path_abs, NAME_CHRONICS_FOLDER))
+    else:
+        # otherwise use it
+        chronics_path_abs = os.path.abspath(chronics_path)
+    _check_path(chronics_path_abs, "Dataset chronics folder")
+
+    # Compute and find backend/grid file
+    grid_path = _get_default_aux("grid_path", kwargs,
+                                 defaultClassApp=str, defaultinstance="",
+                                 msg_error=ERR_MSG_KWARGS["grid_path"])
+    if grid_path == "":
+        grid_path_abs = os.path.abspath(os.path.join(dataset_path_abs, NAME_GRID_FILE))
+    else:
+        grid_path_abs = os.path.abspath(grid_path)
+    _check_path(grid_path_abs, "Dataset power flow solver configuration")
+
+    # Compute and find grid layout file
+    grid_layout_path_abs = os.path.abspath(os.path.join(dataset_path_abs, NAME_GRID_LAYOUT_FILE))
+    _check_path(grid_layout_path_abs, "Dataset grid layout")
+
+    # Check provided config overrides are valid
+    _check_kwargs(kwargs)
+
+    # Compute and find config file
+    config_path_abs = os.path.abspath(os.path.join(dataset_path_abs, NAME_CONFIG_FILE))
+    _check_path(config_path_abs, "Dataset environment configuration")
+
+    # Read config file
+    try:
+        spec = importlib.util.spec_from_file_location("config.config", config_path_abs)
+        config_module = importlib.util.module_from_spec(spec)
+        spec.loader.exec_module(config_module)
+        config_data = config_module.config
+    except Exception as e:
+        print (e)
+        raise EnvError("Invalid dataset config file: {}".format(config_path_abs)) from None
+
+    # Get graph layout
+    try:
+        with open(grid_layout_path_abs) as layout_fp:
+            graph_layout = json.load(layout_fp)
+    except Exception as e:
+        raise EnvError("Dataset {} doesn't have a valid graph layout".format(config_path_abs))
+
+    # Get thermal limits
+    thermal_limits = None
+    if "thermal_limits" in config_data:
+        thermal_limits = config_data["thermal_limits"]
+
+    # Get chronics_to_backend
+    name_converter = None
+    if "names_chronics_to_grid" in config_data:
+        name_converter = config_data["names_chronics_to_grid"]
+    if name_converter is None:
+        name_converter = {}
+    names_chronics_to_backend = _get_default_aux("names_chronics_to_backend", kwargs,
+                                                 defaultClassApp=dict, defaultinstance=name_converter,
+                                                 msg_error=ERR_MSG_KWARGS["names_chronics_to_grid"])
+    # Get default backend class
+    backend_class_cfg = PandaPowerBackend
+    if "backend_class" in config_data and config_data["backend_class"] is not None:
+        backend_class_cfg = config_data["backend_class"]
+    ## Create the backend, to compute the powerflow
+    backend = _get_default_aux("backend", kwargs, defaultClass=backend_class_cfg,
+                               defaultClassApp=Backend, msg_error=ERR_MSG_KWARGS["backend"])
+
+    # Get default observation class
+    observation_class_cfg = CompleteObservation
+    if "observation_class" in config_data and config_data["observation_class"] is not None:
+        observation_class_cfg = config_data["observation_class"]
+    ## Setup the type of observation the agent will receive
+    observation_class = _get_default_aux("observation_class", kwargs, defaultClass=observation_class_cfg, isclass=True,
+                                         defaultClassApp=BaseObservation, msg_error=ERR_MSG_KWARGS["observation_class"])
+
+    ## Create the parameters of the game, thermal limits threshold,
+    # simulate cascading failure, powerflow mode etc. (the gamification of the game)
+    param = _get_default_aux('param', kwargs, defaultClass=Parameters,
+                             defaultClassApp=Parameters, msg_error=ERR_MSG_KWARGS["param"])
+
+    # Get default rules class
+    rules_class_cfg = DefaultRules
+    if "rules_class" in config_data and config_data["rules_class"] is not None:
+        rules_class_cfg = config_data["rules_class"]
+    ## Create the rules of the game (mimic the operationnal constraints)
+    gamerules_class = _get_default_aux("gamerules_class", kwargs, defaultClass=rules_class_cfg,
+                                       defaultClassApp=BaseRules, msg_error=ERR_MSG_KWARGS["gamerules_class"],
+                                       isclass=True)
+
+    # Get default reward class
+    reward_class_cfg = L2RPNReward
+    if "reward_class" in config_data and config_data["reward_class"] is not None:
+        reward_class_cfg = config_data["reward_class"]
+    ## Setup the reward the agent will receive
+    reward_class = _get_default_aux("reward_class", kwargs, defaultClass=reward_class_cfg,
+                                    defaultClassApp=BaseReward, msg_error=ERR_MSG_KWARGS["reward_class"],
+                                    isclass=True)
+
+    # Get default BaseAction class
+    action_class_cfg = BaseAction
+    if "action_class" in config_data and config_data["action_class"] is not None:
+        action_class_cfg = config_data["action_class"]
+    ## Setup the type of action the BaseAgent can perform
+    action_class = _get_default_aux("action_class", kwargs, defaultClass=action_class_cfg,
+                                    defaultClassApp=BaseAction, msg_error=ERR_MSG_KWARGS["action_class"],
+                                    isclass=True)
+
+    # Get default Voltage class
+    voltage_class_cfg = ControlVoltageFromFile
+    if "voltage_class" in config_data and config_data["voltage_class"] is not None:
+        voltage_class_cfg = config_data["voltage_class"]
+    ### Create controler for voltages
+    volagecontroler_class = _get_default_aux("volagecontroler_class", kwargs,
+                                             defaultClassApp=voltage_class_cfg,
+                                             defaultClass=ControlVoltageFromFile,
+                                             msg_error=ERR_MSG_KWARGS["voltagecontroler_class"], isclass=True)
+
+    # Get default Chronics class
+    chronics_class_cfg = ChangeNothing
+    if "chronics_class" in config_data and config_data["chronics_class"] is not None:
+        chronics_class_cfg = config_data["chronics_class"]
+    # Get default Grid class
+    grid_value_class_cfg = GridStateFromFile
+    if "grid_value_class" in config_data and config_data["grid_value_class"] is not None:
+        grid_value_class_cfg = config_data["grid_value_class"]
+
+    ## the chronics to use
+    ### the arguments used to build the data, note that the arguments must be compatible with the chronics class
+    default_chronics_kwargs = {
+        "chronicsClass": chronics_class_cfg,
+        "path": chronics_path_abs,
+        "gridvalueClass": grid_value_class_cfg
+    }
+
+    data_feeding_kwargs = _get_default_aux("data_feeding_kwargs", kwargs,
+                                           defaultClassApp=dict,
+                                           defaultinstance=default_chronics_kwargs,
+                                           msg_error=ERR_MSG_KWARGS["data_feeding_kwargs"])
+    for el in default_chronics_kwargs:
+        if not el in data_feeding_kwargs:
+            data_feeding_kwargs[el] = default_chronics_kwargs[el]
+
+    ### the chronics generator
+    chronics_class_used = _get_default_aux("chronics_class", kwargs,
+                                           defaultClassApp=GridValue,
+                                           defaultClass=data_feeding_kwargs["chronicsClass"],
+                                           msg_error=ERR_MSG_KWARGS["chronics_class"],
+                                           isclass=True)
+    data_feeding_kwargs["chronicsClass"] = chronics_class_used
+    data_feeding = _get_default_aux("data_feeding", kwargs,
+                                    defaultClassApp=ChronicsHandler,
+                                    defaultClass=ChronicsHandler,
+                                    build_kwargs=data_feeding_kwargs,
+                                    msg_error=ERR_MSG_KWARGS["chronics_handler"])
+
+    ### other rewards
+    other_rewards = _get_default_aux("other_rewards", kwargs,
+                                     defaultClassApp=dict,
+                                     defaultinstance={},
+                                     msg_error=ERR_MSG_KWARGS["other_rewards"],
+                                     isclass=False)
+
+    # Opponent
+    # TODO make that in config file of the default environment !!!
+    opponent_action_class = _get_default_aux("opponent_action_class",
+                                             kwargs,
+                                             defaultClassApp=BaseAction,
+                                             defaultClass=DontAct,
+                                             msg_error=ERR_MSG_KWARGS["opponent_action_class"],
+                                             isclass=True)
+    opponent_class = _get_default_aux("opponent_class",
+                                      kwargs,
+                                      defaultClassApp=BaseOpponent,
+                                      defaultClass=BaseOpponent,
+                                      msg_error=ERR_MSG_KWARGS["opponent_class"],
+                                      isclass=True)
+    opponent_init_budget = _get_default_aux("opponent_init_budget", kwargs,
+                                            defaultClassApp=float,
+                                            defaultinstance=0.,
+                                            msg_error=ERR_MSG_KWARGS["opponent_init_budget"],
+                                            isclass=False)
+
+    # Finally instanciate env from config & overrides
+    env = Environment(init_grid_path=grid_path_abs,
+                      chronics_handler=data_feeding,
+                      backend=backend,
+                      parameters=param,
+                      names_chronics_to_backend=names_chronics_to_backend,
+                      actionClass=action_class,
+                      observationClass=observation_class,
+                      rewardClass=reward_class,
+                      legalActClass=gamerules_class,
+                      voltagecontrolerClass=volagecontroler_class,
+                      other_rewards=other_rewards,
+                      opponent_action_class=opponent_action_class,
+                      opponent_class=opponent_class,
+                      opponent_init_budget=opponent_init_budget)
+
+    # Update the thermal limit if any
+    if thermal_limits is not None:
+        env.set_thermal_limit(thermal_limits)
+
+    # Set graph layout if not None and not an empty dict
+    if graph_layout is not None and graph_layout:
+        env.attach_layout(graph_layout)
+
+    return env
