+++ grid2op/MakeEnv/__init__.py
@@ -1,25 +1,27 @@
-__all__ = [ "make_new",
-            "list_available_remote_env",
-            "list_available_local_env",
-            "get_current_local_dir",
-            "change_local_dir",
-            # deprecated in v 0.8.0
-            "make",
-            "make2",
-            # super deprecated
-            "CASE_14_FILE",
-            "CHRONICS_FODLER",
-            "CHRONICS_MLUTIEPISODE",
-            "NAMES_CHRONICS_TO_BACKEND",
-            "EXAMPLE_CHRONICSPATH",
-            "EXAMPLE_CASEFILE",
-            "L2RPN2019_CASEFILE",
-            "L2RPN2019_DICT_NAMES"]
+__all__ = [
+    "make",
+    "list_available_remote_env",
+    "list_available_local_env",
+    "get_current_local_dir",
+    "change_local_dir",
+    # deprecated in v 0.8.0
+    "make",
+    "make_from_dataset_path",
+    # super deprecated
+    "CASE_14_FILE",
+    "CHRONICS_FODLER",
+    "CHRONICS_MLUTIEPISODE",
+    "NAMES_CHRONICS_TO_BACKEND",
+    "EXAMPLE_CHRONICSPATH",
+    "EXAMPLE_CASEFILE",
+    "L2RPN2019_CASEFILE",
+    "L2RPN2019_DICT_NAMES"
+]
 
-from grid2op.MakeEnv.MakeEnv import make_from_dataset_path as make2
-from grid2op.MakeEnv.MakeEnv import make
-from grid2op.MakeEnv.MakeNew import make_new
+from grid2op.MakeEnv.MakeOld import make_old
+from grid2op.MakeEnv.MakeFromPath import make_from_dataset_path
+from grid2op.MakeEnv.Make import make
 from grid2op.MakeEnv.UserUtils import list_available_remote_env, list_available_local_env, get_current_local_dir, change_local_dir
 
-from grid2op.MakeEnv.MakeEnv import CASE_14_FILE, CHRONICS_FODLER, CHRONICS_MLUTIEPISODE, NAMES_CHRONICS_TO_BACKEND
-from grid2op.MakeEnv.MakeEnv import EXAMPLE_CHRONICSPATH, EXAMPLE_CASEFILE, L2RPN2019_CASEFILE, L2RPN2019_DICT_NAMES
+from grid2op.MakeEnv.MakeOld import CASE_14_FILE, CHRONICS_FODLER, CHRONICS_MLUTIEPISODE, NAMES_CHRONICS_TO_BACKEND
+from grid2op.MakeEnv.MakeOld import EXAMPLE_CHRONICSPATH, EXAMPLE_CASEFILE, L2RPN2019_CASEFILE, L2RPN2019_DICT_NAMES
