+++ grid2op/data/blank/grid.json
@@ -1,0 +1,1 @@
+{}
