+++ grid2op/data/blank/grid_layout.json
@@ -1,0 +1,1 @@
+{}
