+++ grid2op/tests/test_Environment.py
@@ -19,7 +19,7 @@ from grid2op.Backend import PandaPowerBackend
 from grid2op.Parameters import Parameters
 from grid2op.Chronics import ChronicsHandler, GridStateFromFile, ChangeNothing
 from grid2op.Reward import L2RPNReward
-from grid2op.MakeEnv import make_new
+from grid2op.MakeEnv import make
 from grid2op.Rules import RulesChecker, DefaultRules
 from grid2op.Action import *
 
@@ -190,7 +190,7 @@ class TestIllegalAmbiguous(unittest.TestCase):
         self.tol_one = 1e-4
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
-            self.env = make_new("rte_case5_example", test=True)
+            self.env = make("rte_case5_example", test=True)
 
     def tearDown(self):
         self.env.close()
@@ -235,7 +235,7 @@ class TestOtherReward(unittest.TestCase):
         self.tol_one = 1e-4
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
-            self.env = make_new("rte_case5_example", test=True, reward_class=L2RPNReward,
+            self.env = make("rte_case5_example", test=True, reward_class=L2RPNReward,
                                 other_rewards={"test": L2RPNReward})
 
     def tearDown(self):
@@ -262,7 +262,7 @@ class TestAttachLayout(unittest.TestCase):
 
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
-            with make_new("rte_case5_example", test=True, reward_class=L2RPNReward, other_rewards={"test": L2RPNReward}) \
+            with make("rte_case5_example", test=True, reward_class=L2RPNReward, other_rewards={"test": L2RPNReward}) \
                     as env:
                 env.attach_layout(my_layout)
                 act = env.action_space()
@@ -297,7 +297,7 @@ class TestLineChangeLastBus(unittest.TestCase):
 
             with warnings.catch_warnings():
                 warnings.filterwarnings("ignore")
-                self.env = make_new("rte_case14_test", test=True, chronics_class=ChangeNothing, param=self.params)
+                self.env = make("rte_case14_test", test=True, chronics_class=ChangeNothing, param=self.params)
 
     def tearDown(self):
         self.env.close()
@@ -375,7 +375,7 @@ class TestResetAfterCascadingFailure(unittest.TestCase):
             warnings.filterwarnings("ignore")
             params = Parameters()
             params.MAX_SUB_CHANGED = 2
-            self.env = make_new("rte_case14_test", test=True, chronics_class=ChangeNothing, param=params)
+            self.env = make("rte_case14_test", test=True, chronics_class=ChangeNothing, param=params)
 
     def tearDown(self):
         self.env.close()
@@ -415,7 +415,7 @@ class TestCascadingFailure(unittest.TestCase):
             params.MAX_SUB_CHANGED = 0
             params.NB_TIMESTEP_POWERFLOW_ALLOWED = 2
             rules = DefaultRules
-            self.env = make_new("rte_case14_test", test=True, chronics_class=ChangeNothing, param=params,
+            self.env = make("rte_case14_test", test=True, chronics_class=ChangeNothing, param=params,
                                 gamerules_class=rules)
 
     def tearDown(self):
