+++ grid2op/tests/test_Runner.py
@@ -15,7 +15,7 @@ PATH_ADN_CHRONICS_FOLDER = os.path.abspath(os.path.join(PATH_CHRONICS, "test_mul
 from grid2op.Chronics import Multifolder
 from grid2op.Reward import L2RPNReward
 from grid2op.Backend import PandaPowerBackend
-from grid2op.MakeEnv import make_new
+from grid2op.MakeEnv import make
 from grid2op.Runner import Runner
 
 
@@ -82,14 +82,14 @@ class TestRunner(HelperTests):
     def test_init_from_env(self):
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
-            with make_new("rte_case14_test", test=True) as env:
+            with make("rte_case14_test", test=True) as env:
                 runner = Runner(**env.get_params_for_runner())
         runner.run(nb_episode=1, max_iter=self.max_iter )
 
     def test_init_from_env_with_other_reward(self):
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
-            with make_new("rte_case14_test", test=True, other_rewards={"test": L2RPNReward}) as env:
+            with make("rte_case14_test", test=True, other_rewards={"test": L2RPNReward}) as env:
                 runner = Runner(**env.get_params_for_runner())
         runner.run(nb_episode=1, max_iter=self.max_iter)
