+++ grid2op/tests/test_VoltageControler.py
@@ -10,14 +10,14 @@ import pdb
 import warnings
 from grid2op.tests.helper_path_test import *
 from grid2op.VoltageControler import ControlVoltageFromFile
-from grid2op.MakeEnv import make_new
+from grid2op.MakeEnv import make
 
 
 class TestLoadingVoltageControl(unittest.TestCase):
     def test_creation_ControlVoltage(self):
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore")
-            with make_new("rte_case5_example", test=True) as env:
+            with make("rte_case5_example", test=True) as env:
                 volt_cont = ControlVoltageFromFile(controler_backend=env.backend, gridobj=env.backend)
